Reorganized IPC code.
