+++ src/app/ipc/DatabaseChannel.ts
@@ -1,9 +1,9 @@
-import { IpcChannelInterface } from "./IpcChannelInterface";
+import { Prisma, PrismaClient, File } from '@prisma/client';
 import { IpcMainEvent } from 'electron';
-import { IpcRequest } from "../../shared/IpcRequest";
-import { Prisma, PrismaClient, File } from '@prisma/client'
+import { IpcRequest } from "./IpcRequest";
+import { IpcChannelInterface } from "./IpcChannelInterface";
 import { Stopwatch } from "../../shared/stopwatch";
-import { Thumbnail } from "../../app/components/virtual-canvas/virtual-canvas.data-source";
+import { Thumbnail } from "../components/virtual-canvas/virtual-canvas.data-source";
 
 export class DatabaseChannel implements IpcChannelInterface {
   readonly db = new PrismaClient();
